+++ packages/components/CHANGELOG.md
@@ -26,6 +26,10 @@
 
 -   `Tabs`: implement new `tabId` prop ([#56883](https://github.com/WordPress/gutenberg/pull/56883)).
 
+### Experimental
+
+-   `Tabs`: improve focus handling in controlled mode ([#56658](https://github.com/WordPress/gutenberg/pull/56658)).
+
 ## 25.13.0 (2023-11-29)
 
 ### Enhancements